+++ commands/Moderation/kick.js
@@ -22,6 +22,21 @@ module.exports = {
           }
         })
       }
+      let allGuilds = client.guilds.cache.array();
+      for (let i = 0; i < allGuilds.length; i++) {
+      Settings.findOne(
+        { guildID: allGuilds[i].id },
+        async (err, settings) => {
+          if (err) console.log(err);
+
+          if (!settings) {
+            enableModeration = false;
+          } else {
+            enableModeration = settings.enableModeration
+          }
+        })
+      }
+      if(enableModeration === true) {
         if (message.deletable) message.delete();
 
         // No args
@@ -106,4 +121,5 @@ module.exports = {
             }
         });
     }
+}
 };
